+++ modules/generator/registry/test.go
@@ -169,5 +169,5 @@ func (t *testHistogram) ObserveWithExemplar(labelValueCombo *LabelValueCombo, va
 func withLe(lbls labels.Labels, le float64) labels.Labels {
 	lb := labels.NewBuilder(lbls)
 	lb.Set(labels.BucketLabel, formatFloat(le))
-	return lb.Labels(nil)
+	return lb.Labels()
 }