+++ modules/generator/storage/config.go
@@ -6,6 +6,7 @@ import (
 
 	prometheus_config "github.com/prometheus/prometheus/config"
 	"github.com/prometheus/prometheus/tsdb/agent"
+	"github.com/prometheus/prometheus/tsdb/wlog"
 )
 
 type Config struct {
@@ -17,6 +18,9 @@ type Config struct {
 	// How long to wait when flushing sample on shutdown
 	RemoteWriteFlushDeadline time.Duration `yaml:"remote_write_flush_deadline"`
 
+	// Add X-Scope-OrgID header in remote write requests
+	RemoteWriteAddOrgIDHeader bool `yaml:"remote_write_add_org_id_header,omitempty"`
+
 	// Prometheus remote write config
 	// https://prometheus.io/docs/prometheus/latest/configuration/configuration/#remote_write
 	RemoteWrite []prometheus_config.RemoteWriteConfig `yaml:"remote_write,omitempty"`
@@ -26,18 +30,20 @@ func (cfg *Config) RegisterFlagsAndApplyDefaults(string, *flag.FlagSet) {
 	cfg.Wal = agentDefaultOptions()
 
 	cfg.RemoteWriteFlushDeadline = time.Minute
+
+	cfg.RemoteWriteAddOrgIDHeader = true
 }
 
 // agentOptions is a copy of agent.Options but with yaml struct tags. Refer to agent.Options for
 // documentation.
 type agentOptions struct {
-	WALSegmentSize    int           `yaml:"wal_segment_size"`
-	WALCompression    bool          `yaml:"wal_compression"`
-	StripeSize        int           `yaml:"stripe_size"`
-	TruncateFrequency time.Duration `yaml:"truncate_frequency"`
-	MinWALTime        int64         `yaml:"min_wal_time"`
-	MaxWALTime        int64         `yaml:"max_wal_time"`
-	NoLockfile        bool          `yaml:"no_lockfile"`
+	WALSegmentSize    int                  `yaml:"wal_segment_size"`
+	WALCompression    wlog.CompressionType `yaml:"wal_compression"`
+	StripeSize        int                  `yaml:"stripe_size"`
+	TruncateFrequency time.Duration        `yaml:"truncate_frequency"`
+	MinWALTime        int64                `yaml:"min_wal_time"`
+	MaxWALTime        int64                `yaml:"max_wal_time"`
+	NoLockfile        bool                 `yaml:"no_lockfile"`
 }
 
 func agentDefaultOptions() agentOptions {